Add code for querying Windows program heaps
This was intended to be an examination of basics/motiv03.cpp from
Josuttis, but got sidetracked in trying to pin down stack and heap
ranges on Windows. Not only does "stack addresses are high and heap
addresses are low" not hold, there isn't even a single heap on Windows.
Turns out there are _multiple_ heaps, all at seemingly random
addresses. Added code sourced from MSDN to print out the number and
addresses of heaps.